refactor: post 도메인 아키텍처 구조 변경
