+++ src/main/java/org/sopt/post/dto/request/UpdatePostRequest.java
@@ -1,4 +1,4 @@
-package org.sopt.dto.request;
+package org.sopt.post.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
