+++ src/main/java/org/sopt/post/dto/response/CreatePostResponse.java
@@ -1,4 +1,4 @@
-package org.sopt.dto.response;
+package org.sopt.post.dto.response;
 
 // 게시글 작성 응답 (서버 → 클라이언트)
 public record CreatePostResponse(
