+++ src/main/java/org/sopt/post/repository/PostRepository.java
@@ -1,7 +1,7 @@
-package org.sopt.repository;
+package org.sopt.post.repository;
 
-import org.sopt.domain.BoardType;
-import org.sopt.domain.Post;
+import org.sopt.post.entity.BoardType;
+import org.sopt.post.entity.Post;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
