+++ src/main/java/org/sopt/post/service/PostService.java
@@ -1,18 +1,17 @@
-package org.sopt.service;
+package org.sopt.post.service;
 
-import org.sopt.domain.BoardType;
-import org.sopt.domain.Post;
-import org.sopt.domain.User;
-import org.sopt.dto.request.CreatePostRequest;
-import org.sopt.dto.request.UpdatePostRequest;
-import org.sopt.dto.response.CreatePostResponse;
-import org.sopt.dto.response.PostResponse;
+import org.sopt.post.entity.BoardType;
+import org.sopt.post.entity.Post;
+import org.sopt.user.entity.User;
+import org.sopt.post.dto.request.CreatePostRequest;
+import org.sopt.post.dto.request.UpdatePostRequest;
+import org.sopt.post.dto.response.CreatePostResponse;
+import org.sopt.post.dto.response.PostResponse;
 import org.sopt.global.code.ErrorCode;
 import org.sopt.global.exception.BaseException;
 import org.sopt.global.exception.PostNotFoundException;
-import org.sopt.repository.PostRepository;
-import org.sopt.repository.UserRepository;
-import org.sopt.validator.PostValidator;
+import org.sopt.post.repository.PostRepository;
+import org.sopt.user.repository.UserRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
